+++ e2e/e2etests/test_sui_withdraw_and_call.go
@@ -12,12 +12,13 @@ import (
 )
 
 func TestSuiWithdrawAndCall(r *runner.E2ERunner, args []string) {
-	require.Len(r, args, 1)
+	require.Len(r, args, 2)
 
 	// ARRANGE
-	// Given target package ID (example package) and a SUI amount
+	// Given target package ID (example package), SUI amount and gas limit
 	targetPackageID := r.SuiExample.PackageID.String()
 	amount := utils.ParseBigInt(r, args[0])
+	gasLimit := utils.ParseBigInt(r, args[1])
 
 	// use the deployer address as on_call payload message
 	signer, err := r.Account.SuiSigner()
@@ -29,18 +30,20 @@ func TestSuiWithdrawAndCall(r *runner.E2ERunner, args []string) {
 	calledCountBefore := r.SuiGetConnectedCalledCount()
 
 	// create the on_call payload
-	payloadOnCall, err := r.SuiCreateExampleWACPayload(suiAddress)
-	require.NoError(r, err)
+	authorizedSender := r.EVMAddress()
+	payloadOnCall := r.SuiCreateExampleWACPayload(authorizedSender, suiAddress)
 
 	// ACT
 	// approve SUI ZRC20 token
 	r.ApproveSUIZRC20(r.GatewayZEVMAddr)
 
-	// perform the withdraw and call
-	tx := r.SuiWithdrawAndCallSUI(
+	// perform the withdraw and authenticated call
+	tx := r.SuiWithdrawAndCall(
 		targetPackageID,
 		amount,
+		r.SUIZRC20Addr,
 		payloadOnCall,
+		gasLimit,
 		gatewayzevm.RevertOptions{OnRevertGasLimit: big.NewInt(0)},
 	)
 	r.Logger.EVMTransaction(tx, "withdraw_and_call")
@@ -48,7 +51,7 @@ func TestSuiWithdrawAndCall(r *runner.E2ERunner, args []string) {
 	// ASSERT
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
-	r.Logger.CCTX(*cctx, "withdraw")
+	r.Logger.CCTX(*cctx, "withdraw_and_call")
 	require.EqualValues(r, crosschaintypes.CctxStatus_OutboundMined, cctx.CctxStatus.Status)
 
 	// balance after