+++ e2e/e2etests/test_sui_withdraw_restricted_address.go
@@ -31,9 +31,10 @@ func TestSuiWithdrawRestrictedAddress(r *runner.E2ERunner, args []string) {
 
 	// ACT
 	// perform the withdraw to restricted receiver
-	tx := r.SuiWithdrawSUI(
+	tx := r.SuiWithdraw(
 		receiver,
 		amount,
+		r.SUIZRC20Addr,
 		gatewayzevm.RevertOptions{
 			RevertAddress:    revertAddress,
 			OnRevertGasLimit: big.NewInt(0),